Don't detach LDAPd

+++ config/finalizers/ldapd.rb
@@ -1,13 +1,6 @@
 include ApplicationHelper
 
-if server?
+if server? and LDAPd.pid
   Rails.logger.info "Stopping LDAPd with PID #{LDAPd.pid}"
-
-  if LDAPd.pid
-    begin
-      Process.kill 'KILL', LDAPd.pid
-    rescue Errno::ESRCH
-      Rails.logger.error "LDAPd with PID #{LDAPd.pid} already reaped"
-    end
-  end
+  Process.kill 'KILL', LDAPd.pid # raises Errno::ESRCH
 end
